MCP-Based Multi-Agent/requirements.txt
fastapi
uvicorn
requests
openai
streamlit
markdown2
python-dotenv

# to install all at once, run:
# pip install -r requirements.txt